🐛 YD5J89 task: repair hosted CI gates
