+++ .agentplane/tasks/202607221908-YD5J89/pr/meta.json
@@ -2,7 +2,7 @@
   "base": "main",
   "branch": "task/202607221908-YD5J89/migrate-context-and-evaluator-command-boundaries",
   "created_at": "2026-08-01T00:38:26.423Z",
-  "diffstat_sha256": "sha256:a2aafd6503bfd0ecffb7e19ed5cfdcd5470d364c1ff5417f87f9b900a421066c",
+  "diffstat_sha256": "sha256:a21515297583fa962d73422d14ae1d5361f88f2f71e424bf74b14f1bac122dd0",
   "last_verified_at": "2026-08-01T01:34:41.204Z",
   "last_verified_diffstat_sha256": "sha256:87b2352aac3906dfdcd927e3123256ddbfb2ae6157e3411ac4b1619d07657e50",
   "pr_number": 4716,
